claude修改
1. src/modules/src/uart5_pos.c — float → double 显式转换：
  - 第 64 行和第 96 行：pos.x → (double)pos.x（同理 y、z）
  - eprintf 是可变参数函数，float 会自动提升为 double，新 GCC 通过 -Werror=double-promotion 将此视为错误
  2. Makefile — GCC 版本兼容：
  - -Wno-stringop-overread 和 -Wno-stringop-overflow 改为条件添加，仅在编译器支持时才启用（旧版 GCC
  不识别这些选项）

+++ src/modules/src/uart5_pos.c
@@ -61,7 +61,7 @@ static void uart5PosTaskFcn(void* param)
     estimatorKalmanGetEstimatedPos(&pos);
 
     // Format: POS:x,y,z\r\n  (example: POS:1.234,-0.567,0.890)
-    uart5Printf("POS:%.3f,%.3f,%.3f\r\n", pos.x, pos.y, pos.z);
+    uart5Printf("POS:%.3f,%.3f,%.3f\r\n", (double)pos.x, (double)pos.y, (double)pos.z);
   }
 }
 
@@ -93,7 +93,7 @@ void uart5SendPosition(void)
   pos.z = 0.0f;
   estimatorKalmanGetEstimatedPos(&pos);
 
-  uart5Printf("POS:%.3f,%.3f,%.3f\r\n", pos.x, pos.y, pos.z);
+  uart5Printf("POS:%.3f,%.3f,%.3f\r\n", (double)pos.x, (double)pos.y, (double)pos.z);
 }
 
 bool uart5PositionSenderTest(void)
